requirements.txt
streamlit
google-generativeai
python-dotenv
poppler-utils
pdf2image
azure-ai-inference 
PyPDF2